+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next Release
 
+#### Breaking Changes
+
+- feat: stop deleting old log entries when a model with the same pk is created (i.e. the pk value is reused) ([#559](https://github.com/jazzband/django-auditlog/pull/559))
+
 #### Improvements
 
 - feat: Added support for Correlation ID